src/main/java/com/fastcampus/newboardproject/service/ArticleService.java
package com.fastcampus.newboardproject.service;

import com.fastcampus.newboardproject.domain.type.SearchType;
import com.fastcampus.newboardproject.dto.ArticleDto;
import com.fastcampus.newboardproject.dto.ArticleUpdateDto;
import com.fastcampus.newboardproject.repository.ArticleRepository;
import java.time.LocalDateTime;
import java.util.List;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Transactional
@Service
public class ArticleService {

    private final ArticleRepository articleRepository;

    @Transactional(readOnly = true)
    public Page<ArticleDto> searchArticles(SearchType title, String search_keyword) {
        return Page.empty();
    }

    @Transactional(readOnly = true)
    public ArticleDto searchArticle(long l) {
        return null;
    }

    public void saveArticle(ArticleDto dto) {
    }

    public void updateArticle(long articleId, ArticleUpdateDto dto) {
    }

    public void deleteArticle(long articleId) {
    }
}
